Refactor: reminder 모듈 생성
- data, domain, presentation layer 분리
- MainActivity.kt, ReminderViewModel.kt 파일 presentation 모듈로 이동
