+++ app/src/main/java/com/minux/reminder/feature_reminder/presentation/MainActivity.kt
@@ -1,7 +1,8 @@
-package com.minux.reminder
+package com.minux.reminder.feature_reminder.presentation
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
+import com.minux.reminder.R
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
